+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
-## [0.0.2]
+## [0.4.0]
+
+- Fix collapse tree functionality
+
+- Add error message when no language server installed
+
+## [0.3.0]
+
+- Fix favourites only working for typescript files
+
+## [0.2.0]
 
 - Update README
 
@@ -8,6 +18,6 @@
 
 - Fix activation events
 
-## [0.0.1]
+## [0.1.0]
 
 - Initial release